requirements.txt
fastapi
uvicorn
sentence-transformers
psycopg2-binary
python-dotenv
pandas
qdrant-client
mysql-connector-python